Prevent adding already existed key
