+++ FlowCrypt/Controllers/Setup/SetupViewController.swift
@@ -15,7 +15,7 @@ final class SetupViewController: ASViewController<ASTableNode> {
     private let imap: Imap
     private let userService: UserServiceType
     private let router: GlobalRouterType
-    private let storage: DataServiceType
+    private let storage: DataServiceType & KeyDataServiceType
     private let decorator: SetupViewDecoratorType
     private let core: Core
     private let keyMethods: KeyMethodsType
@@ -56,7 +56,7 @@ final class SetupViewController: ASViewController<ASTableNode> {
         imap: Imap = Imap.shared,
         userService: UserServiceType = UserService.shared,
         router: GlobalRouterType = GlobalRouter(),
-        storage: DataServiceType = DataService.shared,
+        storage: DataServiceType & KeyDataServiceType = DataService.shared,
         decorator: SetupViewDecoratorType = SetupViewDecorator(),
         core: Core = Core.shared,
         keyMethods: KeyMethodsType = KeyMethods(core: .shared)
