+++ FlowCrypt/Functionality/DataManager/DataService.swift
@@ -16,18 +16,20 @@ protocol DataServiceType {
     var isLoggedIn: Bool { get }
     var isSetupFinished: Bool { get }
     var currentAuthType: AuthType? { get }
-    var keys: [PrvKeyInfo]? { get }
-    var publicKey: String? { get }
 
     // Local data
     var trashFolderPath: String? { get }
     func saveTrashFolder(path: String?)
 
-    func addKeys(keyDetails: [KeyDetails], passPhrase: String, source: KeySource)
-
     // login / logout
     func startFor(user type: SessionType)
     func logOutAndDestroyStorage()
+}
+
+protocol KeyDataServiceType {
+    var keys: [PrvKeyInfo]? { get }
+    var publicKey: String? { get }
+    func addKeys(keyDetails: [KeyDetails], passPhrase: String, source: KeySource)
 }
 
 protocol ImapSessionProvider {
@@ -68,6 +70,20 @@ final class DataService: DataServiceType {
         encryptedStorage.getUser()?.authType
     }
 
+    private let encryptedStorage: EncryptedStorageType & LogOutHandler
+    private(set) var localStorage: LocalStorageType & LogOutHandler
+
+    private init(
+        encryptedStorage: EncryptedStorageType & LogOutHandler = EncryptedStorage(),
+        localStorage: LocalStorageType & LogOutHandler = LocalStorage()
+    ) {
+        self.encryptedStorage = encryptedStorage
+        self.localStorage = localStorage
+    }
+}
+
+// MARK: - DataKeyServiceType
+extension DataService: KeyDataServiceType {
     var keys: [PrvKeyInfo]? {
         guard let keys = encryptedStorage.keys() else { return nil }
         return Array(keys)
@@ -78,15 +94,8 @@ final class DataService: DataServiceType {
         encryptedStorage.publicKey()
     }
 
-    private let encryptedStorage: EncryptedStorageType & LogOutHandler
-    private(set) var localStorage: LocalStorageType & LogOutHandler
-
-    private init(
-        encryptedStorage: EncryptedStorageType & LogOutHandler = EncryptedStorage(),
-        localStorage: LocalStorageType & LogOutHandler = LocalStorage()
-    ) {
-        self.encryptedStorage = encryptedStorage
-        self.localStorage = localStorage
+    func addKeys(keyDetails: [KeyDetails], passPhrase: String, source: KeySource) {
+        encryptedStorage.addKeys(keyDetails: keyDetails, passPhrase: passPhrase, source: source)
     }
 }
 
@@ -207,10 +216,6 @@ extension DataService {
             logOutAndDestroyStorage()
             encryptedStorage.saveUser(with: user)
         }
-    }
-
-    func addKeys(keyDetails: [KeyDetails], passPhrase: String, source: KeySource) {
-        encryptedStorage.addKeys(keyDetails: keyDetails, passPhrase: passPhrase, source: source)
     }
 }
